Renamed all perfect squares to aps and made a py version.

+++ C++/Math/aps.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-	int x = 1;
+	unsigned long long x = 1;
 	while(true)
 	{
 		cout << x*x << endl;
